+++ packages/spacecat-shared-data-access/test/it/db.test.js
@@ -12,9 +12,9 @@
 
 /* eslint-env mocha */
 
-import chai from 'chai';
+import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import DynamoDbLocal from 'dynamo-db-local';
+import { spawn } from 'dynamo-db-local';
 import Joi from 'joi';
 
 import { isIsoDate } from '@adobe/spacecat-shared-utils';
@@ -29,8 +29,7 @@ import { createSiteCandidate, SITE_CANDIDATE_SOURCES, SITE_CANDIDATE_STATUS } fr
 import { KEY_EVENT_TYPES } from '../../src/models/key-event.js';
 import { ConfigurationDto } from '../../src/dto/configuration.js';
 
-const { expect } = chai;
-chai.use(chaiAsPromised);
+use(chaiAsPromised);
 
 function checkSite(site, configuration) {
   expect(site).to.be.an('object');
@@ -100,6 +99,9 @@ const TEST_DA_CONFIG = {
   tableNameConfigurations: 'spacecat-services-configurations',
   tableNameSiteTopPages: 'spacecat-services-site-top-pages',
   tableNameExperiments: 'spacecat-services-experiments',
+  tableNameApiKeys: 'spacecat-services-api-keys',
+  tableNameImportJobs: 'spacecat-services-import-jobs',
+  tableNameImportUrls: 'spacecat-services-import-urls',
   indexNameAllSites: 'spacecat-services-all-sites',
   indexNameAllKeyEventsBySiteId: 'spacecat-services-key-events-by-site-id',
   indexNameAllSitesOrganizations: 'spacecat-services-all-sites-organizations',
@@ -107,6 +109,7 @@ const TEST_DA_CONFIG = {
   indexNameAllOrganizationsByImsOrgId: 'spacecat-services-all-organizations-by-ims-org-id',
   indexNameAllSitesByDeliveryType: 'spacecat-services-all-sites-by-delivery-type',
   indexNameAllLatestAuditScores: 'spacecat-services-all-latest-audit-scores',
+  indexNameAllImportJobsByStatus: 'spacecat-services-all-import-jobs-by-status',
   pkAllSites: 'ALL_SITES',
   pkAllOrganizations: 'ALL_ORGANIZATIONS',
   pkAllLatestAudits: 'ALL_LATEST_AUDITS',
@@ -135,28 +138,75 @@ describe('DynamoDB Integration Test', async () => {
     process.env.AWS_ACCESS_KEY_ID = 'dummy';
     process.env.AWS_SECRET_ACCESS_KEY = 'dummy';
 
-    dynamoDbLocalProcess = DynamoDbLocal.spawn({
+    dynamoDbLocalProcess = spawn({
+      detached: true,
+      stdio: 'inherit',
       port: 8000,
       sharedDb: true,
     });
 
     await sleep(10000); // give db time to start up
 
-    await generateSampleData(
-      TEST_DA_CONFIG,
-      NUMBER_OF_ORGANIZATIONS,
-      NUMBER_OF_SITES,
-      NUMBER_OF_SITES_CANDIDATES,
-      NUMBER_OF_AUDITS_PER_TYPE_AND_SITE,
-      NUMBER_OF_TOP_PAGES_FOR_SITE,
-      NUMBER_OF_KEY_EVENTS_PER_SITE,
-    );
+    try {
+      await generateSampleData(
+        TEST_DA_CONFIG,
+        NUMBER_OF_ORGANIZATIONS,
+        NUMBER_OF_SITES,
+        NUMBER_OF_SITES_CANDIDATES,
+        NUMBER_OF_AUDITS_PER_TYPE_AND_SITE,
+        NUMBER_OF_TOP_PAGES_FOR_SITE,
+        NUMBER_OF_KEY_EVENTS_PER_SITE,
+      );
+    } catch (e) {
+      console.error('Error generating sample data', e);
+    }
 
     dataAccess = createDataAccess(TEST_DA_CONFIG, console);
   });
 
   after(() => {
     dynamoDbLocalProcess.kill();
+  });
+
+  it('get all key events for a site', async () => {
+    const siteId = (await dataAccess.getSites())[0].getId();
+
+    const keyEvents = await dataAccess.getKeyEventsForSite(siteId);
+
+    expect(keyEvents.length).to.equal(NUMBER_OF_KEY_EVENTS_PER_SITE);
+    expect(keyEvents[0].getSiteId()).to.equal(siteId);
+
+    // check if the key events are returned in descending order
+    for (let i = 1; i < keyEvents.length; i += 1) {
+      const prev = keyEvents[i - 1];
+      const next = keyEvents[i];
+      const desc = prev.getCreatedAt() >= next.getCreatedAt();
+      expect(desc).to.be.true;
+    }
+  });
+
+  it('add a new key event for a site', async () => {
+    const siteId = (await dataAccess.getSites())[0].getId();
+
+    await dataAccess.createKeyEvent({
+      siteId,
+      name: 'new-key-event',
+      type: KEY_EVENT_TYPES.CONTENT,
+    });
+
+    const keyEvents = await dataAccess.getKeyEventsForSite(siteId);
+
+    expect(keyEvents.length).to.equal(NUMBER_OF_KEY_EVENTS_PER_SITE + 1);
+  });
+
+  it('remove a key event', async () => {
+    const siteId = (await dataAccess.getSites())[0].getId();
+    const keyEvents = await dataAccess.getKeyEventsForSite(siteId);
+
+    await dataAccess.removeKeyEvent(keyEvents[0].getId());
+
+    const keyEventsAfter = await dataAccess.getKeyEventsForSite(siteId);
+    expect(keyEventsAfter.length).to.equal(NUMBER_OF_KEY_EVENTS_PER_SITE);
   });
 
   it('gets configuration by Version', async () => {
@@ -736,14 +786,6 @@ describe('DynamoDB Integration Test', async () => {
     expect(latestAuditAfterRemoval).to.be.null;
   });
 
-  it('removes organization', async () => {
-    const organizations = await dataAccess.getOrganizations();
-    const organization = organizations[0];
-    await expect(dataAccess.removeOrganization(organization.getId())).to.eventually.be.fulfilled;
-    const organizationAfterRemoval = await dataAccess.getOrganizationByID(organization.getId());
-    expect(organizationAfterRemoval).to.be.null;
-  });
-
   it('verify a previously added site candidate exists', async () => {
     const exists = await dataAccess.siteCandidateExists('https://example0.com');
     expect(exists).to.be.true;
@@ -859,83 +901,47 @@ describe('DynamoDB Integration Test', async () => {
     expect(topPagesAfterRemoval).to.be.an('array').that.is.empty;
   });
 
-  it('get all key events for a site', async () => {
-    const siteId = (await dataAccess.getSites())[0].getId();
-
-    const keyEvents = await dataAccess.getKeyEventsForSite(siteId);
-
-    expect(keyEvents.length).to.equal(NUMBER_OF_KEY_EVENTS_PER_SITE);
-    expect(keyEvents[0].getSiteId()).to.equal(siteId);
-
-    // check if the key events are returned in descending order
-    for (let i = 1; i < keyEvents.length; i += 1) {
-      const prev = keyEvents[i - 1];
-      const next = keyEvents[i];
-      const desc = prev.getCreatedAt() >= next.getCreatedAt();
-      expect(desc).to.be.true;
-    }
-  });
-
-  it('add a new key event for a site', async () => {
-    const siteId = (await dataAccess.getSites())[0].getId();
-
-    await dataAccess.createKeyEvent({
-      siteId,
-      name: 'new-key-event',
-      type: KEY_EVENT_TYPES.CONTENT,
-    });
-
-    const keyEvents = await dataAccess.getKeyEventsForSite(siteId);
-
-    expect(keyEvents.length).to.equal(NUMBER_OF_KEY_EVENTS_PER_SITE + 1);
-  });
-
-  it('remove a key event', async () => {
-    const siteId = (await dataAccess.getSites())[0].getId();
-    const keyEvents = await dataAccess.getKeyEventsForSite(siteId);
-
-    await dataAccess.removeKeyEvent(keyEvents[0].getId());
-
-    const keyEventsAfter = await dataAccess.getKeyEventsForSite(siteId);
-    expect(keyEventsAfter.length).to.equal(NUMBER_OF_KEY_EVENTS_PER_SITE);
-  });
-
   it('get all experiments for the site', async () => {
-    const sites = await dataAccess.getSites();
-    const experiments = await dataAccess.getExperiments(sites[0].getId());
+    const site = await dataAccess.getSiteByBaseURL('https://example0.com');
+    const siteId = site.getId();
+    const experiments = await dataAccess.getExperiments(siteId);
 
     expect(experiments.length).to.equal(NUMBER_OF_EXPERIMENTS);
   });
 
   it('get all experiments for the site and experimentId', async () => {
     // handling multi page experiments
-    const sites = await dataAccess.getSites();
-    const experiments = await dataAccess.getExperiments(sites[0].getId(), 'experiment-1');
+    const site = await dataAccess.getSiteByBaseURL('https://example0.com');
+    const siteId = site.getId();
+    const experiments = await dataAccess.getExperiments(siteId, 'experiment-1');
 
     expect(experiments.length).to.equal(1);
   });
 
   it('get 0 experiments for the siteId with out any experiments', async () => {
-    const sites = await dataAccess.getSites();
-    const experiments = await dataAccess.getExperiments(sites[1].getId());
+    const site = await dataAccess.getSiteByBaseURL('https://example3.com');
+    const siteId = site.getId();
+    const experiments = await dataAccess.getExperiments(siteId);
 
     expect(experiments.length).to.equal(0);
   });
 
   it('check if experiment exists', async () => {
-    const sites = await dataAccess.getSites();
-    const experiment = await dataAccess.getExperiment(sites[0].getId(), 'experiment-1', `${sites[0].getBaseURL()}/page-1`);
+    const site = await dataAccess.getSiteByBaseURL('https://example0.com');
+    const siteId = site.getId();
+    const experiment = await dataAccess.getExperiment(siteId, 'experiment-1', `${site.getBaseURL()}/page-1`);
 
     expect(experiment).to.not.equal(null);
   });
 
   it('create and update experiment', async () => {
-    const sites = await dataAccess.getSites();
+    const site = await dataAccess.getSiteByBaseURL('https://example0.com');
+    const siteId = site.getId();
     const experimentData = {
-      siteId: sites[0].getId(),
+      siteId,
       experimentId: 'experiment-test',
       name: 'Experiment Test',
-      url: `${sites[0].getBaseURL()}/page-10`,
+      url: `${site.getBaseURL()}/page-10`,
       status: 'active',
       type: 'full',
       variants: [
@@ -945,7 +951,7 @@ describe('DynamoDB Integration Test', async () => {
           interactionsCount: 40,
           p_value: 'coming soon',
           split: 0.5,
-          url: `${sites[0].baseURL}/page-10/variant-1`,
+          url: `${site.getBaseURL()}/page-10/variant-1`,
           views: 1100,
           metrics: [
             {
@@ -961,7 +967,7 @@ describe('DynamoDB Integration Test', async () => {
           p_value: 'coming soon',
           metrics: [],
           split: 0.5,
-          url: `${sites[0].baseURL}/page-10`,
+          url: `${site.getBaseURL()}/page-10`,
           views: 1090,
         },
       ],
@@ -971,12 +977,22 @@ describe('DynamoDB Integration Test', async () => {
       updatedBy: 'it-test',
     };
     await dataAccess.upsertExperiment(experimentData);
-    const experimentTest = await dataAccess.getExperiment(sites[0].getId(), 'experiment-test', `${sites[0].getBaseURL()}/page-10`);
+    const experimentTest = await dataAccess.getExperiment(siteId, 'experiment-test', `${site.getBaseURL()}/page-10`);
     expect(experimentTest).to.not.equal(null);
     // update the experiment variant 0 metrics to 50
     experimentData.variants[0].metrics[0].value = 50;
     await dataAccess.upsertExperiment(experimentData);
-    const updatedExperiment = await dataAccess.getExperiment(sites[0].getId(), 'experiment-test', `${sites[0].getBaseURL()}/page-10`);
+    const updatedExperiment = await dataAccess.getExperiment(siteId, 'experiment-test', `${site.getBaseURL()}/page-10`);
     expect(updatedExperiment.getVariants()[0].metrics[0].value).to.equal(50);
+  });
+
+  /** Please keep this test at the last, the remove organization is removing the org randomly,
+   * moving this  above may remove the org which may have sites required by other tests */
+  it('removes organization', async () => {
+    const organizations = await dataAccess.getOrganizations();
+    const organization = organizations[0];
+    await expect(dataAccess.removeOrganization(organization.getId())).to.eventually.be.fulfilled;
+    const organizationAfterRemoval = await dataAccess.getOrganizationByID(organization.getId());
+    expect(organizationAfterRemoval).to.be.null;
   });
 });